+++ R/utils.R
@@ -548,7 +548,7 @@ has_.data <- function(link) {
 build_.data_link <- function(link, node, dag_sm) {
   dep_codes <- unlist(node, use.names = FALSE)
   deps_funcs <- dep_codes[!grepl("^join.", dep_codes)]
-  all_links <- dag_sm$get_links()
+  all_links <- dag_sm$get("links")
   dep_links <- lapply(deps_funcs, \(code) all_links[[code]])
   dep_queries <- get_supplied_queries(links = dep_links)
   link$args <- sapply(link$args, \(arg) {