+++ tests/testthat/test-dependencies.R
@@ -2,7 +2,7 @@ test_that("A query with no dependencies returns the query", {
   expected_query <- query(customer_data = c("name"))
   dag_sm <- build_tree(test_network)
   test_query <- dependencies(query_list = expected_query, state_list = NULL, network = test_network, dag_sm = dag_sm)
-  expect_identical(test_query$get_query(), expected_query)
+  expect_identical(test_query$get("query"), expected_query)
 })
 
 test_that("Query with a dependecy returns the query with the dependency", {
@@ -10,7 +10,7 @@ test_that("Query with a dependecy returns the query with the dependency", {
   dag_sm <- build_tree(test_network)
   test_query <- dependencies(query_list = query, state_list = NULL, network = test_network, dag_sm = dag_sm)
   expected_query <- query(customer_data = c("category", "score", "name"))
-  expect_identical(test_query$get_query(), expected_query)
+  expect_identical(test_query$get("query"), expected_query)
 })
 
 test_that("Query from two layers returns the entire query", {
@@ -18,7 +18,7 @@ test_that("Query from two layers returns the entire query", {
   dag_sm <- build_tree(test_network)
   test_query <- dependencies(query_list = query, state_list = NULL, network = test_network, dag_sm = dag_sm)
   expected_query <- query(customer_data = c("category", "score", "name"))
-  expect_identical(test_query$get_query(), expected_query)
+  expect_identical(test_query$get("query"), expected_query)
 })
 
 test_that("dependencies works with an empty list and returns a named empty query", {
@@ -26,7 +26,7 @@ test_that("dependencies works with an empty list and returns a named empty query
   dag_sm <- build_tree(test_network)
   test_query <- dependencies(query_list =  query, state_list = NULL, network = test_network, dag_sm = dag_sm)
   expected_query <- query()
-  expect_identical(test_query$get_query(), expected_query)
+  expect_identical(test_query$get("query"), expected_query)
 })
 
 test_that("dependencies from two projects can be retrieved and hidden dependencies revealed", {
@@ -36,7 +36,7 @@ test_that("dependencies from two projects can be retrieved and hidden dependenci
   dag_sm$set_main_project("customer_data")
   sm <- resolve_dependencies(query_list = globals$internal, state_list = globals$states, network = test_network, dag_sm = dag_sm)
   expected_query <- query(customer_data = c("category", "score", "name", "id"), occupations = c("department", "id"))
-  expect_identical(sm$get_query(), expected_query)
+  expect_identical(sm$get("query"), expected_query)
 })
 
 test_that("on entry takes copies foreign dependency of nodes", {